Them chuyen dong nhan vat

+++ Object.h
@@ -12,7 +12,7 @@ public:
 	SDL_Rect& getSrc();
 	SDL_Rect& getDest();
 	void setSrc(int x, int y, int w, int h);
-	void setDest(int x, double y, int w, int h);
+	void setDest(float x, float y, int w, int h);
 
 	void CreateTexture(const char* address, SDL_Renderer* ren);
 	void virtual Render(SDL_Renderer* ren) = 0;
